Implement single-cycle frontend simulator
Auto-DSE does not have RF build in for simulation, so I removed the RF
based part and use fatptr as two oprands for ADD and MUL directly.

Ideally there's no problems with this taken. It can still show the
superiorty of HMT, but I don't know if Sandhya and Kevin are fine with
this method.

The next step would be modify the front end into 2-stages(IF/ID) and add
Auto-DSE instruction fusion.