Abstract simulation setup into Sim and add support for SCALE_ADD
The timing model for SCALE_ADD does not really works for instuciton
fusion. This is because SCALE_ADD will not yield READ_SRCx event, and
HMT cycle for now exists in READ_SRCx event.

In order to prevent other instruction be like this, the best way is to
implement a HMT event which delay the execution for certain cycles. But
for single instruction benchmarks(vec-add, axpy), it's fine to stick
    with the old timing model.